requirements.txt
asgiref==3.6.0
Django==3.2.18
pip==21.3.1
pytz==2023.3
setuptools==60.2.0
sqlparse==0.4.3
typing-extensions==4.5.0
wheel==0.37.1
python-dotenv~=0.21.1
djangorestframework==3.12.4
psycopg2-binary==2.9.6
djoser==2.1.0
django-filter==21.1
Pillow==9.5.0
